프로토타입 모듈화
