+++ oop/data.js
@@ -1,0 +1,6 @@
+module.exports = [
+    { 'name': 'asdf1', 'tag': [ 'asdf' ], 'status': 'todo', 'id': 9547 },
+    { 'name': 'asdf2', 'tag': [ 'asdf' ], 'status': 'todo', 'id': 9546 },
+    { 'name': 'asdf3', 'tag': [ 'asdf' ], 'status': 'todo', 'id': 9545 },
+    { 'name': 'asdf4', 'tag': [ 'asdf' ], 'status': 'todo', 'id': 9544 }
+];
